Consolidate mentor forms
